+++ game/client/beamdraw.h
@@ -131,31 +131,31 @@ public:
 int ScreenTransform( const Vector& point, Vector& screen );
 
 void DrawSegs( int noise_divisions, float *prgNoise, const model_t* spritemodel,
-				float frame, int rendermode, const Vector& source, const Vector& delta, 
+				int frame, int rendermode, const Vector& source, const Vector& delta, 
 				float startWidth, float endWidth, float scale, float freq, float speed, int segments,
 				int flags, float* color, float fadeLength, float flHDRColorScale = 1.0f );
 void DrawTeslaSegs( int noise_divisions, float *prgNoise, const model_t* spritemodel,
-				float frame, int rendermode, const Vector& source, const Vector& delta, 
+				int frame, int rendermode, const Vector& source, const Vector& delta, 
 				float startWidth, float endWidth, float scale, float freq, float speed, int segments,
 				int flags, float* color, float fadeLength, float flHDRColorScale = 1.0f );
 void DrawSplineSegs( int noise_divisions, float *prgNoise, 
 				const model_t* beammodel, const model_t* halomodel, float flHaloScale,
-				float frame, int rendermode, int numAttachments, Vector* attachment, 
+				int frame, int rendermode, int numAttachments, Vector* attachment, 
 				float startWidth, float endWidth, float scale, float freq, float speed, int segments,
 				int flags, float* color, float fadeLength, float flHDRColorScale = 1.0f );
 void DrawHalo(IMaterial* pMaterial, const Vector& source, float scale, float const* color, float flHDRColorScale = 1.0f );
-void BeamDrawHalo( const model_t* spritemodel, float frame, int rendermode, const Vector& source, 
+void BeamDrawHalo( const model_t* spritemodel, int frame, int rendermode, const Vector& source, 
 				  float scale, float* color, float flHDRColorScale = 1.0f );
 void DrawDisk( int noise_divisions, float *prgNoise, const model_t* spritemodel,
-			  float frame, int rendermode, const Vector& source, const Vector& delta, 
+			  int frame, int rendermode, const Vector& source, const Vector& delta, 
 			  float width, float scale, float freq, float speed, 
 			  int segments, float* color, float flHDRColorScale = 1.0f );
 void DrawCylinder( int noise_divisions, float *prgNoise, const model_t* spritemodel, 
-				  float frame, int rendermode, const Vector& source, 
+				  int frame, int rendermode, const Vector& source, 
 				  const Vector&  delta, float width, float scale, float freq, 
 				  float speed, int segments, float* color, float flHDRColorScale = 1.0f );
 void DrawRing( int noise_divisions, float *prgNoise, void (*pfnNoise)( float *noise, int divs, float scale ), 
-			  const model_t* spritemodel, float frame, int rendermode, 
+			  const model_t* spritemodel, int frame, int rendermode, 
 			  const Vector& source, const Vector& delta, float width, float amplitude, 
 			  float freq, float speed, int segments, float* color, float flHDRColorScale = 1.0f );
 void DrawBeamFollow( const model_t* spritemodel, BeamTrail_t* pHead, int frame, int rendermode, Vector& delta,
